update original init
